requirements.txt
absl-py==2.0.0
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
ansi2html==1.9.1
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.2
av==11.0.0
Babel==2.14.0
beautifulsoup4==4.12.2
bidict==0.22.1
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.0
comet-ml==3.35.5
comm==0.2.0
ConfigArgParse==1.7
configobj==5.0.8
cryptography==41.0.7
cycler==0.12.1
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
defusedxml==0.7.1
descartes==1.1.0
docker-pycreds==0.4.0
docstring-parser==0.15
dulwich==0.21.7
embreex==2.17.7.post4
everett==3.1.0
exceptiongroup==1.2.0
fastjsonschema==2.19.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.0
fonttools==4.47.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
google-auth-oauthlib==1.0.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.26.0
icecream==2.1.3
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.0
importlib-resources==6.1.1
ipykernel==6.27.1
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jaxtyping==0.2.19
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.10.0
lit==17.0.6
lpips==0.1.4
lxml==4.9.4
mapbox-earcut==1.0.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.5.3
mdurl==0.1.2
mediapy==1.2.0
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.7
msgpack-numpy==0.4.8
multidict==6.0.4
nbclient==0.9.0
nbconvert==7.13.1
nbformat==5.7.0
nerfacc==0.5.2
nest-asyncio==1.5.8
networkx==3.1
ninja==1.11.1.1
nodeenv==1.8.0
notebook_shim==0.2.3
numpy==1.24.4
nuscenes-devkit==1.1.11
nvidia-ml-py==11.525.112
nvitop==1.1.2
oauthlib==3.2.2
open3d==0.17.0
opencv-python==4.6.0.66
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
plotly==5.18.0
prometheus-client==0.19.0
protobuf==3.20.3
psutil==5.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycodestyle==2.10.0
pycollada==0.7.2
pycparser==2.21
pyliblzfse==0.4.1
pymeshlab==2023.12
pyngrok==7.0.3
pyparsing==3.1.1
pypng==0.20220715.0
pyquaternion==0.9.9
PySocks==1.7.1
python-box==6.1.0
python-dateutil==2.8.2
python-engineio==4.8.0
python-json-logger==2.0.7
python-socketio==5.10.0
pytorch-lightning==2.1.3
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.2
rawpy==0.19.0
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.15.2
rsa==4.9
Rtree==1.1.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentry-sdk==1.39.1
setproctitle==1.3.3
Shapely==1.8.5.post1
shtab==1.6.5
simple-websocket==1.0.0
simplejson==3.19.2
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
splines==0.3.0
svg.path==6.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.3.0
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.6.7
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
tqdm==4.65.0
trimesh==4.0.7
triton==2.0.0
typeguard==4.1.5
types-python-dateutil==2.8.19.14
tyro==0.6.1
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
viser==0.1.17
wandb==0.16.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.3.3
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.16.0
wsproto==1.2.0
wurlitzer==3.0.3
xatlas==0.0.8
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
yourdfpy==0.0.53
zipp==3.17.0
